requirements.txt
streamlit
torch
Pillow
numpy
